fix: move GPTQWeight into file to avoid circular import
